schedule core

+++ boom-job-common/src/main/java/me/stevenkin/boom/job/common/model/Job.java
@@ -18,7 +18,7 @@ public class Job {
     private LocalDateTime startTime;
     private LocalDateTime endTime;
     private Integer repeatCount;
-    private Long repeatInterval;
+    private Integer repeatInterval;
     private boolean isRepeatForever;
     private String desc;
     private Long appId;
